requirements.txt
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.6
dateparser==1.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
numpy==1.20.1
pandas==1.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-binance==0.7.9
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
service-identity==18.1.0
six==1.15.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.3
zope.interface==5.2.0
